All intermediate sizes can be cropped to fit
Even if the thumbnail size is the only one than can have the crop configured via the backoffice, the other sizes (medium, medium_large and large) can also be configured to crop via the update_option function. If the latter happens, this change will correct the reported size and crop message.